requirements.txt
Django==2.2
djangorestframework==3.12.4
Jinja2==2.10.1
django-debug-toolbar==1.10.1